Resolved field definition conflicts:
- Adopted i8 for grid_head_len (feature branch)
- Kept i16 for ref_pp_field (main branch)
- Kept i16 for body counts (both)
- Added genome_hint_allocation (feature branch)

+++ gear_optimizer/solver/taichi_gem/api/timeline.py
@@ -148,9 +148,9 @@ def _upload_timeline_grid(timeline_grid):
     _t_extract = time.perf_counter()
 
     # Allocate 3D arrays matching slotted grid fields (slot 0 for CPU upload path)
-    cbf_np = np.zeros((MAX_SONG_SLOTS, GRID_SIZE, GRID_SIZE), dtype=np.int32)
-    cbn_np = np.zeros((MAX_SONG_SLOTS, GRID_SIZE, GRID_SIZE), dtype=np.int32)
-    hl_np = np.zeros((MAX_SONG_SLOTS, GRID_SIZE, GRID_SIZE), dtype=np.int32)
+    cbf_np = np.zeros((MAX_SONG_SLOTS, GRID_SIZE, GRID_SIZE), dtype=np.int16)
+    cbn_np = np.zeros((MAX_SONG_SLOTS, GRID_SIZE, GRID_SIZE), dtype=np.int16)
+    hl_np = np.zeros((MAX_SONG_SLOTS, GRID_SIZE, GRID_SIZE), dtype=np.int8)
     masks_np = np.zeros((MAX_SONG_SLOTS, GRID_SIZE, GRID_SIZE, MAX_HEAD_NOTES), dtype=np.int8)
     masks_bits_np = np.zeros((MAX_SONG_SLOTS, GRID_SIZE, GRID_SIZE, 4), dtype=np.uint32)
